Started conversion of torrent controller
- Many check to be done
- Still some methods to check
Right now it tries to access the cache of the torrent and only in case
of a cache miss tries to update it.  When operating directly on torrents
it avoids to use the cache at all.
The cache main purpose is just to avoid asking to rtorrent if  certain
torrent exists. Maybe a better way should be to check the xmlrpc calls
for errors